[nasa/nos3#463] Started reorganization of CSS checkout. Used old CSS checkout files for now, but appear to be copies of sample, so may go back to current sample when rewriting
