+++ fsw/cfs/CMakeLists.txt
@@ -6,6 +6,7 @@ include(../../../ComponentSettings.cmake)
 include_directories(public_inc)
 include_directories(mission_inc)
 include_directories(platform_inc)
+include_directories(shared)
 include_directories(src)
 
 include_directories(${hwlib_MISSION_DIR}/fsw/public_inc)
