+++ fsw/standalone/device_cfg.h
